+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 5.x.x
 
 - Breaking: Dropped support for TYPO3 11 LTS and PHP < 8.2
+- Breaking: Introduced OidcConfiguration class to represent extension configuration.
+  Default values of endpoints have been removed. Please validate your configuration during upgrade.
 - Feature: Added extension setting `enablePasswordCredentials` to disable password-authentication.
 
 ## Version 4.0.0